[fix] - updates prices to decimal values

+++ server/models/auctionModel.js
@@ -12,13 +12,13 @@ const Auction = connection.define('auction', {
     type: Sequelize.INTEGER
   },
   startPrice: {
-    type: Sequelize.INTEGER
+    type: Sequelize.DECIMAL(6, 2)
   },
   currentPrice: {
-    type: Sequelize.INTEGER
+    type: Sequelize.DECIMAL(6, 2)
   },
   minPrice: {
-    type: Sequelize.INTEGER
+    type: Sequelize.DECIMAL(6, 2)
   },
   numTickets: {
     type: Sequelize.INTEGER
@@ -37,7 +37,7 @@ const Auction = connection.define('auction', {
     type: Sequelize.STRING
   },
   tickets: {
-    type: Sequelize.TEXT
+     type: Sequelize.TEXT
   },
 });
